fastapi_server/requirements.txt
python-dotenv
fastapi
uvicorn
tritonclient[all]
soundfile
numpy
scipy
pydub
shortuuid
azure-storage-blob
